third_party/blink/web_tests/platform/linux-chrome/external/wpt/html/cross-origin-opener-policy/reporting/document-reporting/report-only-four-reports.https-expected.txt

This is a testharness.js-based test.
[FAIL] coop document reporting test A test with both COOP and COOP report only setup using Reporting-Endpoints header to CROSS_ORIGIN with same-origin-allow-popups; report-to="coop-popup-report-endpoint", require-corp, same-origin; report-to="coop-popup-report-only-endpoint", require-corp
  assert_equals: has exactly one report for coop-report-endpoint expected 1 but got 0
Harness: the test ran to completion.
